Restructure tests a little
