+++ application/src/test/java/org/opentripplanner/graph_builder/module/osm/moduletests/ElevatorTest.java
@@ -1,0 +1,67 @@
+package org.opentripplanner.graph_builder.module.osm.moduletests;
+
+import static com.google.common.truth.Truth.assertThat;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.opentripplanner.graph_builder.module.osm.moduletests._support.NodeBuilder.node;
+
+import org.junit.jupiter.api.Test;
+import org.opentripplanner.framework.geometry.WgsCoordinate;
+import org.opentripplanner.graph_builder.module.osm.OsmModule;
+import org.opentripplanner.graph_builder.module.osm.moduletests._support.TestOsmProvider;
+import org.opentripplanner.osm.model.OsmWay;
+import org.opentripplanner.routing.graph.Graph;
+import org.opentripplanner.service.osminfo.internal.DefaultOsmInfoGraphBuildRepository;
+import org.opentripplanner.service.vehicleparking.internal.DefaultVehicleParkingRepository;
+import org.opentripplanner.street.model.edge.ElevatorHopEdge;
+import org.opentripplanner.transit.model.framework.Deduplicator;
+
+class ElevatorTest {
+
+  @Test
+  void testDuration() {
+    var way = new OsmWay();
+    way.addTag("duration", "00:01:02");
+    way.addTag("highway", "elevator");
+    var provider = TestOsmProvider.of().addWay(way).build();
+    var graph = new Graph(new Deduplicator());
+    var osmModule = OsmModule.of(
+      provider,
+      graph,
+      new DefaultOsmInfoGraphBuildRepository(),
+      new DefaultVehicleParkingRepository()
+    ).build();
+    osmModule.buildGraph();
+    var edges = graph.getEdgesOfType(ElevatorHopEdge.class);
+    assertThat(edges).hasSize(2);
+    for (var edge : edges) {
+      assertEquals(62, edge.getTravelTime());
+    }
+  }
+
+  @Test
+  void testMultilevelNodeDuration() {
+    var node0 = node(0, new WgsCoordinate(0, 0));
+    var node1 = node(1, new WgsCoordinate(2, 0));
+    var node = node(2, new WgsCoordinate(1, 0));
+    node.addTag("duration", "00:01:02");
+    node.addTag("highway", "elevator");
+    node.addTag("level", "1;2");
+    var provider = TestOsmProvider.of()
+      .addWayFromNodes(way -> way.addTag("level", "1"), node0, node)
+      .addWayFromNodes(way -> way.addTag("level", "2"), node1, node)
+      .build();
+    var graph = new Graph(new Deduplicator());
+    var osmModule = OsmModule.of(
+      provider,
+      graph,
+      new DefaultOsmInfoGraphBuildRepository(),
+      new DefaultVehicleParkingRepository()
+    ).build();
+    osmModule.buildGraph();
+    var edges = graph.getEdgesOfType(ElevatorHopEdge.class);
+    assertThat(edges).hasSize(2);
+    for (var edge : edges) {
+      assertEquals(62, edge.getTravelTime());
+    }
+  }
+}
