+++ application/src/test/java/org/opentripplanner/graph_builder/module/osm/moduletests/walkablearea/SimpleAreaTest.java
@@ -16,7 +16,7 @@ import org.opentripplanner.street.model.edge.AreaEdge;
 import org.opentripplanner.test.support.GeoJsonIo;
 import org.opentripplanner.transit.model.framework.Deduplicator;
 
-public class SimpleAreaTest {
+class SimpleAreaTest {
 
   @Test
   void walkableArea() {
